projects/ai-coach/requirements.txt
# API Claude
anthropic

# Base vectorielle
chromadb

# Chargement de documents
pypdf
python-docx

# Configuration
python-dotenv

# Interface CLI
rich

# Optionnel: chunking avancé
langchain
langchain-text-splitters
